# openapi.yaml
openapi: 3.1.0
info:
  title: Pocket Concierge API
  version: 0.4.0
  description: Governed household errands and customer-funded Ethereum NFT mint delivery with verifiable receipts.
servers:
  - url: http://localhost:4310
security:
  - bearerAuth: []
paths:
...
  /v1/errands:
    post:
      operationId: createErrand
      summary: Create one idempotent household errand cycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErrandInput"
      responses:
        "201":
          description: Errand created with an authority decision and one next action
        "200":
          description: Idempotent replay of the exact same errand cycle
        "409":
          description: externalId and cycleId already belong to a different immutable errand
...
  /v1/authority/receipts/{receiptId}:
    get:
      security: []
      operationId: getAuthorityReceipt
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: string
            pattern: "^pr_[a-f0-9]{32}$"
      responses:
        "200":
          description: Public privacy-safe authority and execution receipt with recomputed hash verification
        "404":
          description: Receipt not found
  /v1/missions/preview:
    post:
      operationId: previewMission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionInput"
      responses:
        "200":
          description: Deterministic mission preview; nothing is persisted or paid
  /v1/missions:
    post:
      operationId: createMission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionInput"
      responses:
        "201":
          description: Created mission, or an idempotent replay of the same immutable manifest
        "409":
          description: externalId was already used for a different manifest
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    orderToken:
      type: apiKey
      in: header
      name: X-Order-Token
    operatorKey:
      type: apiKey
      in: header
      name: X-Operator-Key
    pilotKey:
      type: apiKey
      in: header
      name: X-Pocket-Pilot-Key
...
  schemas:
...
    ErrandInput:
      type: object
      additionalProperties: false
      required: [externalId, cycleId, dueAt, errand, mandate]
      properties:
        externalId:
          type: string
          description: Stable caller or platform job identifier
        cycleId:
          type: string
          description: Stable one-off, weekly, or monthly cycle identifier
        title: { type: string }
        timezone: { type: string, default: UTC }
        dueAt: { type: string, format: date-time }
        errand:
          type: object
          additionalProperties: false
          required: [category, serviceId, variationCode, privateInputRef, maximumUsdt]
          properties:
            category:
              type: string
              enum: [data, electricity, tv]
            serviceId: { type: string }
            variationCode: { type: string }
            privateInputRef:
              type: string
              description: Opaque key into the buyer agent's local vault
            maximumUsdt: { type: string }
            description: { type: string }
        mandate:
          type: object
          additionalProperties: false
          required: [validFrom, expiresAt]
          properties:
            validFrom: { type: string, format: date-time }
            expiresAt: { type: string, format: date-time }
            maximumPerActionUsdt: { type: string }
            maximumCycleUsdt: { type: string }
            approvalThresholdUsdt: { type: string }
    MissionInput:
      type: object
      additionalProperties: false
      required: [externalId, title, timezone, mandate, actions]
      properties:
        externalId:
          type: string
          minLength: 8
          maxLength: 128
          description: Caller-provided idempotency identifier
        title:
          type: string
          maxLength: 160
        timezone:
          type: string
          examples: [Africa/Lagos]
        mandate:
          $ref: "#/components/schemas/Mandate"
        actions:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: "#/components/schemas/OkxBillAction"
    OkxBillAction:
      type: object
      additionalProperties: false
      required: [type, reference, description, dueAt, category, serviceId, variationCode, privateInputRef, maximumUsdt]
      properties:
        type:
          const: okx_bill
        reference:
          type: string
          description: Unique action reference inside the mission
        description:
          type: string
        dueAt:
          type: string
          format: date-time
        category:
          type: string
          enum: [data, electricity, tv]
        serviceId:
          type: string
        variationCode:
          type: string
        privateInputRef:
          type: string
          description: Opaque local reference; never a phone, meter, or smartcard number
        maximumUsdt:
          type: string
          pattern: "^(?:0|[1-9][0-9]*)(?:\\.[0-9]{1,6})?$"
          description: Hard buyer-side payment ceiling
    Mandate:
      type: object
      additionalProperties: false
      required:
        - policyVersion
        - validFrom
        - expiresAt
        - allowedCategories
        - allowedServiceIds
        - allowedPrivateInputRefs
        - maximumPerActionUsdt
        - maximumMissionUsdt
        - approvalThresholdUsdt
        - maximumActions
      properties:
        policyVersion:
          const: "1"
        validFrom:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        allowedCategories:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            enum: [data, electricity, tv]
        allowedServiceIds:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
        allowedPrivateInputRefs:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
          description: Opaque local recipient references, never raw customer identifiers
        maximumPerActionUsdt:
          type: string
          pattern: "^(?:0|[1-9][0-9]*)(?:\\.[0-9]{1,6})?$"
        maximumMissionUsdt:
          type: string
          pattern: "^(?:0|[1-9][0-9]*)(?:\\.[0-9]{1,6})?$"
        approvalThresholdUsdt:
          type: string
          pattern: "^(?:0|[1-9][0-9]*)(?:\\.[0-9]{1,6})?$"
        maximumActions:
          type: integer
          minimum: 1
          maximum: 10
...